Improve communication protocol.
Use an AtomicI8 as flag (and fix issue with alighnment)
The New flag sets -1 to exit. And every process (client or server)
notifies with its id.
The notification checks that the current flag value is correct
otherwise extra exit methods are now implemented.
On error ani side can notify with a negative value in order to allow
the other side to exit instead of waiti forever. Previous code only
checked the flag once. And overrided it unconditionally.
The new code is also more safe because it checks that only the
expected counserside wrote in the shared buffer.

Stats don't need anymore the enable-disable trick because the client don't
send anymore an empty matrix.

The send function has been generalized in order to send any of the
current Matrix containers. And in the server thread we use the send
function again instead of the trick to update inplace.